requirements.txt.txt
requests
beautifulsoup4
pandas
openpyxl
pillow
rpaframework
selenium
